Move classes to private API
Moves membergroup and SplitBrainMergeTypeProvider classes to private API
since they are not meant to be used by the user.

com.hazelcast.partition.membergroup.BackupSafeMemberGroupFactory
com.hazelcast.partition.membergroup.ConfigMemberGroupFactory
com.hazelcast.partition.membergroup.DefaultMemberGroup
com.hazelcast.partition.membergroup.HostAwareMemberGroupFactory
com.hazelcast.partition.membergroup.MemberGroup
com.hazelcast.partition.membergroup.MemberGroupFactory
com.hazelcast.partition.membergroup.MemberGroupFactoryFactory
com.hazelcast.partition.membergroup.SingleMemberGroup
com.hazelcast.partition.membergroup.SingleMemberGroupFactory
com.hazelcast.partition.membergroup.SPIAwareMemberGroupFactory
com.hazelcast.partition.membergroup.ZoneAwareMemberGroupFactory.java

com.hazelcast.spi.merge.SplitBrainMergeTypeProvider
