+++ hazelcast/src/main/java/com/hazelcast/internal/partition/membergroup/SPIAwareMemberGroupFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.partition.membergroup;
+package com.hazelcast.internal.partition.membergroup;
 
 
 import com.hazelcast.config.InvalidConfigurationException;
