+++ hazelcast/src/test/java/com/hazelcast/internal/partition/impl/PartitionStateGeneratorTest.java
@@ -28,12 +28,12 @@ import com.hazelcast.internal.partition.PartitionReplica;
 import com.hazelcast.internal.partition.PartitionStateGenerator;
 import com.hazelcast.internal.util.UuidUtil;
 import com.hazelcast.cluster.Address;
-import com.hazelcast.partition.membergroup.ConfigMemberGroupFactory;
-import com.hazelcast.partition.membergroup.DefaultMemberGroup;
-import com.hazelcast.partition.membergroup.HostAwareMemberGroupFactory;
-import com.hazelcast.partition.membergroup.MemberGroup;
-import com.hazelcast.partition.membergroup.MemberGroupFactory;
-import com.hazelcast.partition.membergroup.SingleMemberGroupFactory;
+import com.hazelcast.internal.partition.membergroup.ConfigMemberGroupFactory;
+import com.hazelcast.internal.partition.membergroup.DefaultMemberGroup;
+import com.hazelcast.internal.partition.membergroup.HostAwareMemberGroupFactory;
+import com.hazelcast.internal.partition.membergroup.MemberGroup;
+import com.hazelcast.internal.partition.membergroup.MemberGroupFactory;
+import com.hazelcast.internal.partition.membergroup.SingleMemberGroupFactory;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
